refactor: switch token count types from f64 to u64 across provider chain

+++ crates/agent-core-v2/src/agent/llm_requester/contract.rs
@@ -43,7 +43,7 @@ pub struct AgentLlmRequestOverrides {
     pub tools: Option<Vec<Tool>>,
     pub system_prompt: Option<String>,
     pub source: Option<AgentLlmRequestSource>,
-    pub max_output_size: Option<f64>,
+    pub max_output_size: Option<u64>,
 }
 pub struct AgentLlmRequestTask {
     pub trace: LlmRequestTrace,
