+++ crates/agent-core-v2/src/kosong/contract/generate.rs
@@ -456,7 +456,7 @@ mod tests {
             None
         }
 
-        fn max_completion_tokens(&self) -> Option<f64> {
+        fn max_completion_tokens(&self) -> Option<u64> {
             None
         }
